+++ tools/registry.go
@@ -8,6 +8,7 @@ import (
 	"github.com/wt68/runcode/tools/grep"
 	"github.com/wt68/runcode/tools/read"
 	"github.com/wt68/runcode/tools/todo"
+	"github.com/wt68/runcode/tools/webfetch"
 	"github.com/wt68/runcode/tools/write"
 )
 
@@ -20,5 +21,6 @@ func Builtins() []tool.Tool {
 		grep.New(),
 		bash.New(),
 		todo.New(),
+		webfetch.New(),
 	}
 }